\name{plot.correlation}
\alias{plot.correlation}
\title{Plot a correlation matrix}
\usage{
\method{plot}{correlation}(x, y=NULL, ...)
}
\arguments{
    \item{x}{A result of \link{correlation} function;}
    \item{y}{Not used;}
    \item{...}{Not used.}
}
\description{
    Plot a correlation matrix.
}
\details{
    Plot \code{obj}, result of \link{correlation}.
}

%\references{
%    References
%}

\author{Diego Pasqualin \email{dpasqualin@inf.ufpr.br}}
\seealso{
    \link{correlation}
}
%\keyword{phylogenetics}
\keyword{utilities}
