\name{plot}
\alias{plot}
\title{Plot a correlation matrix}
\usage{
plot(obj)
}
\arguments{
    \item{obj}{A result of \link{sfreemap.correlation} function;}
}
\description{
    Plot a correlation matrix.
}
\details{
    Plot \code{obj}, result of \link{sfreemap.correlation} and \link{add.data}.
}

%\references{
%    References
%}

\author{Diego Pasqualin \email{dpasqualin@inf.ufpr.br}}
\seealso{
    \link{sfreemap.correlation}, \link{add.data}
}
%\keyword{phylogenetics}
\keyword{utilities}
