\name{sfreemap-package}
\alias{sfreemap-package}
\docType{package}
\title{
\packageTitle{sfreemap}
}
\description{
\packageDescription{sfreemap}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sfreemap}
\packageIndices{sfreemap}

This package can be used to perform a stochastic mapping in one or more phylogenetic trees, providing as result the expected number of transitions of states and the dwelling times for each branch of the tree.

Afterwards it is possible to analyse the results through charts and plots from tools provided in this package.

The method implemented is based on an article written by Vladimir Minin and Marc Suchard, describing a way to perform analytical stochastic mapping, without requiring to simulations as most methods do. As a result of this, the execution of most functions is really fast.

}
\author{
\packageAuthor{sfreemap}

Maintainer: \packageMaintainer{sfreemap}
}
\references{
Vladimir N Minin e Marc A Suchard. Fast, accurate and simulation-free stochastic mapping. Philosophical Transactions of the Royal Society B: Biological Sciences, 363(1512):3985-3995, 2008.
}
\keyword{ package }
%\seealso{
% Optional links to other man pages, e.g.
% \code{\link[<pkg>:<pkg>-package]{<pkg>}}
%}
%\examples{
%
%}
