\name{sfreemap.corals.tips}
\docType{data}
\alias{sfreemap.corals.tips}
\title{Tip data for \link{sfreemap.corals.trees}}
\description{
One binary character with states "solitary" "colonial").
Complete description on \link{sfreemap.corals.trees}
}
\usage{sfreemap.corals.tips}
\format{A named vector containing the states as values and species as names}
\references{Barbeitos, Marcos S., Sandra L. Romano, and Howard R. Lasker. "Repeated Loss of Coloniality and Symbiosis in Scleractinian Corals." Proceedings of the National Academy of Sciences 107, no. 26 (2010): 11877-82.}
\keyword{datasets}
