\name{sfreemap.primates.dna.tips}
\docType{data}
\alias{sfreemap.primates.dna.tips}
\title{Tip data for \link{sfreemap.primates.trees}}
\description{
Complete description on \link{sfreemap.primates.trees}
}
\usage{sfreemap.primates.dna.tips}
\format{A matrix containing the species as row names and characters as columns}
\keyword{datasets}
