\name{sfreemap.tips.dna}
\docType{data}
\alias{sfreemap.tips.dna}
\title{nucleotide tips for the 901 phylogenetic trees from \link{sfreemap.trees}}
\description{
Description of the data set
}
\usage{sfreemap.tips.dna}
\format{A matrix..}
\source{Marcos Barbeitos (reference).}
\references{
Quote article or something..
}
\keyword{datasets}
