\name{sfreemap.tips}
\docType{data}
\alias{sfreemap.tips}
\title{tips for the 302 phylogenetic trees in sfreemap.trees}
\description{
Description of the data set
}
\usage{sfreemap.tips}
\format{A matrix..}
\source{Marcos Barbeitos (reference).}
\references{
Quote article or something..
}
\keyword{datasets}
