\name{sfreemapc-package}
\alias{sfreemapc-package}
\alias{sfreemapc}
\docType{package}
\title{
\packageTitle{sfreemapc}
}
\description{
\packageDescription{sfreemapc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sfreemapc}
\packageIndices{sfreemapc}

This package can be used to run a stochastic mapping in onde or more phylogenetic
trees, providing as result the expected number of transitions of states and the
dwelling times for each branch of the tree.

The method implemented is based on an article written by Vladimir Minin and Marc
Suchard, describing a way to perform analytical stochastic mapping, without
requiring to simulations, as usual methods do.

}
\author{
\packageAuthor{sfreemapc}

Maintainer: \packageMaintainer{sfreemapc}
}
\references{
Vladimir N Minin e Marc A Suchard. Fast, accurate and simulation-free stochastic
mapping. Philosophical Transactions of the Royal Society B: Biological Sciences,
363(1512):3985-3995, 2008.
}
\keyword{ package }
%\seealso{
% Optional links to other man pages, e.g.
% \code{\link[<pkg>:<pkg>-package]{<pkg>}}
%}
%\examples{
%
%}
