\name{sfreemapc-package}
\alias{sfreemapc-package}
\alias{sfreemapc}
\docType{package}
\title{
\packageTitle{sfreemapc}
}
\description{
\packageDescription{sfreemapc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sfreemapc}
\packageIndices{sfreemapc}
An overview of how to use the package, including the most important
functions
}
\author{
\packageAuthor{sfreemapc}

Maintainer: \packageMaintainer{sfreemapc}
}
\references{
 Literature or other references for background information
}
\keyword{ package }
\seealso{
 Optional links to other man pages, e.g.
 \code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
%\examples{
%
%}
